Move email forwarding to email module
