+++ assets/stylesheets/_components.scss
@@ -318,6 +318,7 @@
 @import 'notices/style';
 @import 'my-sites/domains/domain-management/style';
 @import 'my-sites/domains/domain-management/email/style';
+@import 'my-sites/email/email-forwarding/style';
 @import 'my-sites/domains/domain-management/components/email-verification/style';
 @import 'my-sites/domains/domain-management/components/inbound-transfer-verification/style';
 @import 'my-sites/domains/domain-management/components/icann-verification/style';
