+++ client/my-sites/domains/domain-management/domain-management.jsx
@@ -9,7 +9,6 @@ import DomainConnectMapping from './domain-connect-mapping';
 import Edit from './edit';
 import EditContactInfo from './edit-contact-info';
 import Email from './email';
-import EmailForwarding from './email-forwarding';
 import List from './list';
 import ManageConsent from './manage-consent';
 import NameServers from './name-servers';
@@ -29,7 +28,6 @@ export default {
 	EditContactInfo,
 	ManageConsent,
 	Email,
-	EmailForwarding,
 	List,
 	NameServers,
 	PrimaryDomain,
